+++ src/types/taskFilterType.ts
@@ -1,0 +1,13 @@
+/*** The type for task filters ***/
+
+type taskFilterType = {
+  text: string;
+  priority: {
+    none: boolean;
+    low: boolean;
+    medium: boolean;
+    high: boolean;
+  };
+};
+
+export default taskFilterType;