+++ src/utils/functions/getEmptyTask.ts
@@ -1,4 +1,4 @@
-/*** Return an empty, fresh task of type taskType. ***/
+/*** Returns an empty, fresh task of type taskType. ***/
 
 import { v4 as uuidv4 } from "uuid";
 import taskType from "../../types/taskType";
@@ -9,7 +9,7 @@ const getEmptyTask = (): taskType => {
     name: "",
     category: "",
     description: "",
-    priority: 0,
+    priority: "none",
     isComplete: false,
   };
 };